requirements.txt
# Core dependencies for ai-dev-kit CLI tool
pyyaml>=6.0

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0
# pytest-cov>=4.0
# black>=23.0
# flake8>=6.0
# mypy>=1.0

